+++ src/main/resources/static/css/interviewRealEnter.css
@@ -82,7 +82,7 @@ h2{
     gap: 71px;
     padding: 78px 0 112px 0;
 }
-a{
+.click{
     padding: 15px 28px;
     color: white;
     background-color: #0F4584;
