+++ src/main/resources/templates/layout/main.html
@@ -14,7 +14,7 @@
   <section class="hero information">
     <h1 th:text="'모의 기술 면접 서비스, ' + ${siteName}">모의 기술 면접 서비스, JOB SHIN</h1>
     <p th:text="${heroMessage}">AI가 질문을 하고 피드백까지 제공하는 모의면접을 체험하라!</p>
-    <a href="#" class="btn" th:href="@{/get-started}">Get started</a>
+    <a href="#" class="btn" th:href="@{/views/interviewMainPage}">Get started</a>
   </section>
 
   <section class="feature">
